% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{choisir_difficulte}
\alias{choisir_difficulte}
\title{Calcule le nombre de cases à préremplir selon le niveau de difficulté}
\usage{
choisir_difficulte(taille, niveau)
}
\arguments{
\item{taille}{Un entier représentant la taille de la grille}

\item{niveau}{Le niveau de difficulté : "Débutant", "Amateur" ou "Expert"}
}
\value{
Le nombre de cases à préremplir dans la grille
}
\description{
Calcule le nombre de cases à préremplir selon le niveau de difficulté
}
