% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_theme.R
\name{colour_theme}
\alias{colour_theme}
\title{Obtenir les thèmes de couleur disponibles pour le jeu Takuzu}
\usage{
colour_theme()
}
\value{
Une liste de thèmes avec leurs identifiants, couleurs primaires, accents et contours
}
\description{
Obtenir les thèmes de couleur disponibles pour le jeu Takuzu
}
