% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{donner_indice}
\alias{donner_indice}
\title{Donne un indice basé sur l'état actuel de la grille Takuzu}
\usage{
donner_indice(grille)
}
\arguments{
\item{grille}{Une matrice contenant des 0, 1 et NA}
}
\value{
Une liste contenant un type d'indice et un message d'aide
}
\description{
Donne un indice basé sur l'état actuel de la grille Takuzu
}
