% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{generer_takuzu}
\alias{generer_takuzu}
\title{Génère une grille Takuzu valide avec un niveau donné}
\usage{
generer_takuzu(taille, niveau)
}
\arguments{
\item{taille}{Taille de la grille}

\item{niveau}{Niveau de difficulté : "Débutant", "Amateur", "Expert"}
}
\value{
Une matrice avec des 0, 1 et des NA pour les cases à remplir
}
\description{
Génère une grille Takuzu valide avec un niveau donné
}
