% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isValidLine}
\alias{isValidLine}
\title{Vérifie si une ligne ou colonne de Takuzu respecte les règles du jeu}
\usage{
isValidLine(line)
}
\arguments{
\item{line}{Vecteur contenant les valeurs (0, 1) d'une ligne ou colonne}
}
\value{
Booléen indiquant si la ligne ou colonne est valide selon les règles
}
\description{
Cette fonction vérifie qu'une ligne ou colonne respecte les deux règles principales du Takuzu :
\enumerate{
\item Un nombre égal de 0 et de 1
\item Pas plus de deux chiffres identiques consécutifs
}
}
\examples{
\dontrun{
# Ligne valide avec 3 zeros et 3 uns, sans plus de 2 chiffres identiques consécutifs
isValidLine(c(0, 1, 0, 1, 0, 1))  # TRUE

# Ligne invalide car elle contient trois 1 consécutifs
isValidLine(c(0, 0, 1, 1, 1, 0))  # FALSE
}

}
\keyword{internal}
