% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{verifier_takuzu}
\alias{verifier_takuzu}
\title{Vérifie si une grille Takuzu est correctement remplie}
\usage{
verifier_takuzu(grille)
}
\arguments{
\item{grille}{Une matrice remplie de 0 et 1 (sans NA)}
}
\value{
TRUE si la grille est valide, FALSE sinon
}
\description{
Vérifie si une grille Takuzu est correctement remplie
}
